# swagger.yaml
openapi: 3.0.3
info:
  title: Aether AI Gateway
  version: 1.0.0
  description: 'HTTP API for user, session, and improvement plan report workflows in the Aether core gateway. Every response carries an X-Request-Id header: the identifier of that request in the hub_metrics base.'
servers:
  - url: /
...
paths:
...
  /aether-api/v1/ai/report:
    post:
      tags:
        - Reports
      summary: Analyze an inventory and return the structured extraction
      description: Accepts inventory Markdown and the catalogs in force, runs them through the Aeko analyst flow with the current plan for that inventory as context, stores the new plan, and returns the structured inventory for Postgres persistence.
      operationId: analyzeInventoryReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequestData'
            example:
              id_external_context_inventory: 502
              inventory: '## Escopo 1'
              id_external_user: 12345
              gases:
                - id: 1
                  name: CO2
              scopes:
                - id: 1
                  name: Escopo 1
              categories:
                - id: 1
                  name: Combustão estacionária
                  classification: null
      responses:
        '200':
          description: Structured inventory extracted from the analysis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractedInventoryData'
              example:
                description: Boiler-dominated inventory
                start_period: '2025-01-01'
                end_period: '2025-12-31'
                emissions:
                  - quantity_co2e: 12400.0
                    methodology_description: stationary combustion
                    supplier_data_percentage: null
                    gas: 1
                    scope: 1
                    category: null
                    is_upstream: null
                    is_reduction: false
        '400':
          description: The inventory Markdown is empty or a domain catalog rule was rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: inventory is required to analyze an inventory.
        '404':
          description: No user exists for the supplied external identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: User with id_external_user 12345 not found.
        '502':
          description: The analysis produced no plan or structured inventory in the expected shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: The analysis produced no plan in the shape a report is stored in.
        '503':
          description: Database connection is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Database is not initialized
        '500':
          description: The Aeko SDK is not initialized or an unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: 'Error processing report: ...'
  /aether-api/v1/ai/report/{id_external_inventory}:
    get:
      tags:
        - Reports
      summary: Get the textual improvement plan for an inventory
      description: Returns the stored improvement plan for the external inventory identifier.
      operationId: getInventoryImprovementPlan
      parameters:
        - name: id_external_inventory
          in: path
          required: true
          description: External inventory identifier whose plan should be returned.
          schema:
            type: integer
          example: 502
      responses:
        '200':
          description: Textual improvement plan found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImprovementPlanReportData'
              example:
                defined_problem: high scope 1 emissions
                solving_method: replace the boiler fleet
                reasoning: direct combustion dominates the inventory
        '404':
          description: No improvement plan exists for the supplied inventory identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Improvement plan with id_external_inventory 502 not found.
        '503':
          description: Database connection is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Database is not initialized
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: 'Error retrieving report: ...'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
...
    ReportRequestData:
      type: object
      properties:
        id_external_context_inventory:
          type: integer
          description: External inventory identifier analyzed and stored on the resulting plan.
          example: 502
        inventory:
          type: string
          description: Inventory content as Markdown.
          example: '## Escopo 1'
        id_external_user:
          type: integer
          description: External user identifier responsible for the report.
          example: 12345
        gases:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemData'
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemData'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryCatalogItemData'
      required:
        - id_external_context_inventory
        - inventory
        - id_external_user
        - gases
        - scopes
        - categories
    CatalogItemData:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: CO2
      required:
        - id
        - name
    CategoryCatalogItemData:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Combustão estacionária
        classification:
          type: string
          nullable: true
          enum:
            - UPSTREAM
            - DOWNSTREAM
            - null
          example: null
      required:
        - id
        - name
    ExtractedInventoryData:
      type: object
      properties:
        description:
          type: string
          nullable: true
          description: Short description of the analyzed inventory.
          example: Boiler-dominated inventory
        start_period:
          type: string
          nullable: true
          description: Start of the inventory period.
          example: '2025-01-01'
        end_period:
          type: string
          nullable: true
          description: End of the inventory period.
          example: '2025-12-31'
        emissions:
          type: array
          items:
            $ref: '#/components/schemas/InventoryEmissionData'
      required:
        - description
        - start_period
        - end_period
        - emissions
    InventoryEmissionData:
      type: object
      properties:
        quantity_co2e:
          type: number
          format: float
          example: 12400.0
        methodology_description:
          type: string
          nullable: true
          example: stationary combustion
        supplier_data_percentage:
          type: number
          format: float
          nullable: true
          example: null
        gas:
          type: integer
          nullable: true
          example: 1
        scope:
          type: integer
          nullable: true
          example: 1
        category:
          type: integer
          nullable: true
          example: null
        is_upstream:
          type: boolean
          nullable: true
          example: null
        is_reduction:
          type: boolean
          example: false
      required:
        - quantity_co2e
        - methodology_description
        - supplier_data_percentage
        - gas
        - scope
        - category
        - is_upstream
        - is_reduction
    ImprovementPlanReportData:
      type: object
      properties:
        defined_problem:
          type: string
          description: Problem the analysis identified.
          example: high scope 1 emissions
        solving_method:
          type: string
          description: What the plan proposes doing about it.
          example: replace the boiler fleet
        reasoning:
          type: string
          description: Why that method addresses that problem.
          example: direct combustion dominates the inventory
      required:
        - defined_problem
        - solving_method
        - reasoning
...